+++ src/db_building/build_database.py
@@ -3,7 +3,11 @@ import logging
 import pandas as pd
 import s3fs
 from chromadb.config import Settings
-from config import (
+# from evaluation import RetrievalConfiguration
+from langchain_community.embeddings import HuggingFaceEmbeddings
+from langchain_community.vectorstores import Chroma
+
+from src.config import (
     COLLECTION_NAME,
     DB_DIR_LOCAL,
     EMB_DEVICE,
@@ -11,11 +15,6 @@ from config import (
     MARKDOWN_SEPARATORS,
     S3_BUCKET,
 )
-
-# from evaluation import RetrievalConfiguration
-from langchain_community.embeddings import HuggingFaceEmbeddings
-from langchain_community.vectorstores import Chroma
-
 from .document_chunker import chunk_documents
 from .utils_db import parse_xmls, split_list
