+++ src/utils/__init__.py
@@ -1,4 +1,3 @@
-from .formatting_utilities import format_docs
 from .loading_utilities import load_dataframe_from_parquet
 
-__all__ = ["load_dataframe_from_parquet", "format_docs", "cache_model_from_hf_hub"]
+__all__ = ["load_dataframe_from_parquet", "cache_model_from_hf_hub"]
